requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
google-cloud-storage
requests
python-dotenv
